+++ templates/warlock/src/app/users/models/user/migrations/11-12-2025_23-58-03-user.migration.ts
@@ -7,5 +7,6 @@ export default Migration.create(User, {
   password: string(255),
   image: string(500).nullable(),
   imageMetadata: json().nullable(),
+  lastLogin: timestamp().nullable(),
   deletedAt: timestamp().nullable(),
 });